Create directory for each component, moved styled components to seprate files
