+++ client/src/components/Breadcrumb/style.ts
@@ -1,0 +1,8 @@
+import { Grid } from "@mui/material";
+import { styled } from "@mui/system";
+
+const RootGrid = styled(Grid)({
+  margin: 2,
+});
+
+export { RootGrid };
