+++ client/src/components/CompaniesTableFooter/index.tsx
@@ -1,11 +1,11 @@
 import { Grid } from "@mui/material";
 import { GridSlotsComponentsProps } from "@mui/x-data-grid";
 import Divider from "@mui/material/Divider";
-import UploadFileButton from "./UploadFileButton";
-import FooterPagination from "./TableFooterPagination";
-import ViewButton from "./ViewButton";
+import UploadFileButton from "../UploadButton";
+import FooterPagination from "../TableFooterPagination";
+import ViewButton from "../ViewButton";
 
-function CompanyTableFooter(
+function CompaniesTableFooter(
   props: NonNullable<GridSlotsComponentsProps["footer"]>
 ) {
   return (
@@ -25,4 +25,4 @@ function CompanyTableFooter(
   );
 }
 
-export default CompanyTableFooter;
+export default CompaniesTableFooter;
