+++ client/src/components/Navbar/style.ts
@@ -1,0 +1,13 @@
+import Box from "@mui/material/Box";
+import Typography from "@mui/material/Typography";
+import { styled } from "@mui/system";
+
+const NavbarBox = styled(Box)({
+  flexGrow: 1,
+});
+
+const NavbarTypography = styled(Typography)({
+  flexGrow: 1,
+});
+
+export { NavbarBox, NavbarTypography };
