+++ client/src/components/ReportsTableFooter/index.tsx
@@ -2,9 +2,9 @@ import { Grid } from "@mui/material";
 import { GridSlotsComponentsProps } from "@mui/x-data-grid";
 import Divider from "@mui/material/Divider";
 import { useLocation } from "react-router-dom";
-import UploadFileButton from "./UploadFileButton";
-import FooterPagination from "./TableFooterPagination";
-import ViewButton from "./ViewButton";
+import UploadFileButton from "../UploadButton";
+import FooterPagination from "../TableFooterPagination";
+import ViewButton from "../ViewButton";
 
 function ReportTableFooter(
   props: NonNullable<GridSlotsComponentsProps["footer"]>
