+++ client/src/components/TableFooterPagination/style.ts
@@ -1,0 +1,8 @@
+import { Box } from "@mui/material";
+import { styled } from "@mui/system";
+
+const FooterPaginationBox = styled(Box)({
+  justifyContent: "flex-end",
+});
+
+export { FooterPaginationBox };
