+++ client/src/components/ThemeSwitch/index.tsx
@@ -1,0 +1,7 @@
+import { StyledSwitch } from "./style";
+
+function ThemeSwitch({ theme }: { theme: string }) {
+  return <StyledSwitch checked={theme != "light"} />;
+}
+
+export default ThemeSwitch;
