+++ client/src/components/ViewButton/style.ts
@@ -1,0 +1,8 @@
+import Button from "@mui/material/Button";
+import { styled } from "@mui/system";
+
+const StyledButton = styled(Button)({
+  margin: 10,
+});
+
+export { StyledButton };
